+++ devtools/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.0.3] devtools
 ### Releases
 - [NPM v0.0.3] https://www.npmjs.com/package/@avalanche-sdk/devtools/v/0.0.3 - devtools
+
+## 2025-04-05 12:32:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.529.1 (2.566.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.4] devtools
+### Releases
+- [NPM v0.0.4] https://www.npmjs.com/package/@avalanche-sdk/devtools/v/0.0.4 - devtools